Fix position-0 timestamp leak + rollout dtype mismatch
processor.py: in add_subject, the static-events loop's loop variable
leaks the patient's most recent clinical event into position 0's
per_subject_ages/timestamps. Through rotary pos embeddings, this is
a future-dependent signal correlated with downstream labels. Anchor
position 0 to birth instead.

transformer.py: cast scores_k to float32 before assigning into the
float32 top_scores buffer; the bf16 autocast inference path produced
bf16 scores and triggered a dtype-mismatch RuntimeError.

+++ src/femr/models/transformer.py
@@ -475,7 +475,7 @@ def compute_attention_rollout(
 
             out_idx = torch.where(seg_mask)[0]
             top_positions[out_idx, :k] = pos_k.cpu()
-            top_scores[out_idx, :k] = scores_k.cpu()
+            top_scores[out_idx, :k] = scores_k.float().cpu()
 
         seg_start = seg_end
